+++ settings/example_process_w_imports.yml
@@ -1,5 +1,5 @@
 processing_functions:
-  - function: biahub.analysis.imgproc.binning_czyx
+  - function: biahub.process_data.binning_czyx
     channel: "BF"
     kwargs:
       binning_factor_zyx: [1, 2, 2]
